requirements.txt
numpy==1.10.1
pandas==0.17.0
PyYAML==3.11
scipy==0.16.0
progressbar33==2.4
